refactor: Reworked bash wrapper
Redrafted bash wrapper for invocation of underlying Perl script
awsConn.pl to allow command line arguments for server and
instance arguments.  With use of OPTIND, the input options array
is now emptied/re-initialized with each invocation in a given
session, allowing for the script to be conveniently sourced as
part of a bash profile (for utility functions, and the like).